Refactor: Move as_precondition from _stbt.core to _stbt.precondition
I'm moving non-GStreamer stuff out of _stbt.core.
